refactor checkin components #18
