+++ src/components/pages/index.js
@@ -1,5 +1,5 @@
 import Appointment from "./Appointment";
-import CheckIn from "./CheckIn";
+import CheckIn from "./CheckIn/CheckIn";
 import Welcome from "./Welcome";
 
 export { Appointment, CheckIn, Welcome };
